+++ gdal/port/cpl_vsil_subfile.cpp
@@ -165,7 +165,7 @@ size_t VSISubFileHandle::Read( void * pBuffer, size_t nSize, size_t nCount )
     size_t nByteToRead = nSize * nCount;
     if (nCurOffset + nByteToRead > nSubregionOffset + nSubregionSize)
     {
-        int nRead = (int)VSIFReadL( pBuffer, 1, nSubregionOffset + nSubregionSize - nCurOffset, fp);
+        int nRead = (int)VSIFReadL( pBuffer, 1, (size_t)(nSubregionOffset + nSubregionSize - nCurOffset), fp);
         return nRead / nSize;
     }
     else
@@ -192,7 +192,7 @@ size_t VSISubFileHandle::Write( const void * pBuffer, size_t nSize, size_t nCoun
     size_t nByteToWrite = nSize * nCount;
     if (nCurOffset + nByteToWrite > nSubregionOffset + nSubregionSize)
     {
-        int nWritten = (int)VSIFWriteL( pBuffer, 1, nSubregionOffset + nSubregionSize - nCurOffset, fp);
+        int nWritten = (int)VSIFWriteL( pBuffer, 1, (size_t)(nSubregionOffset + nSubregionSize - nCurOffset), fp);
         return nWritten / nSize;
     }
     else
